// frontend/src/data/sdk/storage/storage.gen.ts
export type storageGetBoxesResponse200ApplicationJson = {
  data: BoxDTO[];
  status: 200;
};